Awesome updates
Finally done the v1 of the bot.
the code is not even clean, but i think is now done to run for my purposes
He gave me a hard work to separate correctly the data from the numbers. I have to keep the link of each user talking to the bot. If an user send a message at the same time of other, their link will concatenate.
The solution i found is to create a database to save, in principle, the users telephone like an unique id  and the link.